:hammer: Dando uma melhorada no DELETE

+++ src/services/DeleteTransactionService.ts
@@ -13,10 +13,16 @@ class DeleteTransactionService {
   public async execute({ id }: RequestService): Promise<DeleteResult> {
     const transactionRepository = getCustomRepository(TransactionsRepository);
 
-    const checkId = validate(id);
+    const validateId = validate(id);
+
+    if (!validateId) {
+      throw new AppError('This ID is not valid.');
+    }
+
+    const checkId = await transactionRepository.findOne(id);
 
     if (!checkId) {
-      throw new AppError('This ID does not exist');
+      throw new AppError('This ID does not exist.');
     }
 
     const transactionDelete = transactionRepository.delete(id);
